check for data available - relible data
the barometer oversamples internally by a rate of 128. THis limits the
output rate to about 2Hz... This will need to be adjusted. However only
this way we get the 1.5Pa resolution.

So I need to weigh speed versus (resolution & stability).

The oversampling removes most noise.

The barometer is polled to wait for a finished acquisition and then
data is only send when its value has been changed. - CAREFUL for later
with multiple sensors!!!